fix(site): address deep review findings
- Extract RoleOption to shared module (modules/users/RoleOption.tsx)
  to eliminate duplication between EditRolesButton and EditUserForm.
- Use role slug as HTML id instead of display name with spaces.
- Fix CollapsibleSummary auto-open: use useState initializer instead
  of useEffect so the advanced section opens correctly on mount.
- Fix oidcRoleSyncEnabled loading flash: default to false when
  deployment config has not loaded yet.
- Simplify canSetRoles conditional.
- Fix WithRoles story to show checked checkboxes.
- Revert unnecessary comment reformatting.
